requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cx_Freeze==7.0.0
docutils==0.21.2
exceptiongroup==1.2.1
filelock==3.14.0
h11==0.14.0
idna==3.7
Kivy==2.3.0
Kivy-Garden==0.1.5
outcome==1.3.0.post0
patchelf==0.17.2.1
pillow==10.3.0
Pygments==2.17.2
PySocks==1.7.1
requests==2.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
